Try using generics
It doesn't work for an undiagnosed reason. It seems like the weak
pointers end up pointing to the wrong object. Need to do more
investigating to see if this is possible at all but something gets
copied. Perhaps if we make the weak pointers and then return an object
to be used it will work.